[Add] 문의 관련 entity,controller,service,repository파일 추가
