+++ backend/src/main/java/com/synergy/backend/question/model/response/ReplyResponseDTO.java
@@ -1,2 +1,16 @@
-package com.synergy.backend.question.model.response;public class ReplyResponseDTO {
+package com.synergy.backend.question.model.response;
+
+import lombok.*;
+
+//요청된 답글 response구조
+
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReplyResponseDTO {
+    private Long atelierIdx;
+    private Long memberIdx;
+    private String ateliername;
+    private String profileImageUrl;
+    private String content;
 }
